Title: 최솟값 3개, Time: 0ms, Memory: 0MB, Status: Compile Error - Codetree

+++ 260422/최솟값 3개/top-3-smallest-number.java
@@ -16,9 +16,9 @@ public class Main {
 
             if(i > 1){
                 if(max == Integer.MAX_VALUE || max > num) {
-                    int a = pq.poll();
-                    int b = pq.poll();
-                    int c = pq.poll();
+                    long a = pq.poll();
+                    long b = pq.poll();
+                    long c = pq.poll();
                     max = c;
                     result = a*b*c;
                     pq.add(a);
